+++ mpv.net/Misc/Misc.cs
@@ -388,6 +388,8 @@ namespace mpvnet
 
     public class PathHelp
     {
+        public static string StartupPath { get; } = Application.StartupPath + "\\";
+
         public static string GetFileName(string path)
         {
             if (string.IsNullOrEmpty(path)) return "";
@@ -396,6 +398,11 @@ namespace mpvnet
             index = path.LastIndexOf('/');
             if (index > -1) return path.Substring(index + 1);
             return path;
+        }
+
+        public static string GetShortExtension(string path)
+        {
+            return Path.GetExtension(path).ToLower().TrimStart('.');
         }
     }
 }