+++ Changelog.md
@@ -15,6 +15,11 @@
   environment variable and the OS default apps settings can be opened (Win 10 only)
 - Error messages are shown when unknown scripts and extensions are found in the startup folder
   because user scripts and extensions are supposed to be located in the config folder instead
+- when the config folder and the startup folder were identical then extensions were loaded
+  twice and script were loaded four times because the script host is an extension, there is
+  now a check to ensure extensions and scripts are only loaded once. On first start there
+  is no longer an option to select the startup folder as config folder, it's still possible
+  using a custom folder but discouraged
 
 ### 5.0